feat: Complete AdvancedMenu functionality
Made a significant enhancement of the AdvancedMenu
working. It now contains all the functionality
that was requested, such as enabling the user to
quickly switch photo provider, or set the output
photos number or to set the inital response page,
from which to extract photos. The latest of those
seem to be not clearly working as expected, so
that it will be further getting some more detailed
insight.