fix: Add a fallback for not found URLs in response
This quick update ensures that no error is thrown
when the query response object does not contain
any sort of link (for one or more photos within
the response object). Instead of causing the app
to crash, some photos which are not provided with
URLs, will simply not be shown.