feat: add a signin / verify / login functionality
This is one of the most extensive updates so far.
The update introduce a custom sigin and login
layout with complete functionality. On top of
that, there is an extra verification step which
involves sending a verification code to the
given e-mail address. This took a lot of testing
to be done correctly, but here it is. The only
part left to implement is the "Reset password"
functionality, but that will be done soon.