fix: keep input text when toggling signin & login
A quick follow-up that negates an issue, where the
user in has some text already added in inputs and
then switch to the another component (login or
signin). Now, even though the component is changed,
if user comes back again to the initial component,
the inputs data will still be there (instead of
being cleared out).