fix: remove input text on form close
This works partially as expected, but still it is
a handy change. Now, when you close a form with
inputs by clicking the FormClose special button,
the input text is cleared. When opening the
component once again, the text is not present
anymore. NOTE: This does not remove input values
for the inputs whose component got closed by a
top-right button