feat: add session / auth functionality
Finished the authentication part of the app and
also provided a session, which keeps track of the
logged user. Whenever user comes back to the
website, he is logged back in because of the
session functionality. User can also manually
log out, which turns off the current session with
an option to reactivate upon next successful login
attempt.