feat: redesign routing system
The routing pattern has been now redefined not to
further rely on state values to set up a given
component. The new version is almost fully based
on URL routing system, which now also beings an
option to define an in-route parameters where
necessary. One of the routes is now supporting
optional parametrs. However, there are still
routes to be prepared and that is the main goal
for the next few updates.