fix: restore integrity for UI and call parameters
This small update patches the recently added
commit. It generally fixes the mismatch between
two type of variables that were used to represent
the actual API call parameter and their input UI
representation. The issue was that when accessing
URL with valid optional parameters compared to the
same search but directly from UI gave different
photo output. Now it is fully fixed and working
as expected.