feat: apply linking to photo author page
The actual photo profile pages on their respective
photo provider sites will be getting more attention
in this update. Now profile pages are more exposed
to the app user to visit, encouraging them to see
some more photos by a visited author. The change
impacts the PhotoPanel author image, and a
individual photo page "link" button, which both
are linked with the author page on click.