chore: refactor fetch functions across components
All the function that are fetch-based (and are
used in more than one component) have been moved
to their respective files, that are responsible
for performing their intended task.

Also there is a small, yet satisfying change for
/home route. That is, every visit fetches
different photos data (as opposed to the previous
state, where only photos of "garden" were
displayed).