fix: apply better loading mechanism for photos
This is a very essential fix. Photos got a new,
much more eye-pleasent onload animation (fade in).
Also optimized images in a way that prevent sudden
Content shift (especially visible on detailed
photo page before). The loading is a bit smoother.
Probably the next update will be a slight
follow-up for the changes.