feat: add a fallback UI for empty response pages
This update includes any kind of extra UI layer,
whenever a page that is supposed to render photos
/ collections, returns an empty response instead.
Typicaly this is because of no actual photos or
collections to be shown. These scenarios are now
covered better - by providing a meaningful
fallback information. Also such pages that might
be using new fallbacks are now better dealing with
scrollbars depending on the situation.