chore: enable photo results to be cached
Added a caching mechanism for photos that has
been already viewed by a user. This change
increased the photos load time in some cases, and
also it reduced redundant API calls. That is true
especially if user searches back for photos that
were already presented before in current view
session.