fix: Prevent UI flicker when adding cached photos
The recently added changes unfortunately
introduced a visual bug. The bug occured when
user changed the page, that contained cached
photos. The sudden UI update was causing a
layout shift, which caused the flickering issue.
Added a "timer" that gives a bit of breathing
room before loading the cached photos, so that
sudden layout shift is no longer an issue.