chore: apply a shadow to panels icons
The previous update cause a little UI issue, that
whenever a dark colored photo was underneath any
of the panel, icons inside the panel were barely
visible. This has been patched by adding a small,
white outline to the icons.

Also, a little change has been made to the Photo
description (inside Photo Panel). The font size
of the description is a bit lighter compared to
the previous version.