chore: prepare auto-fit grid mechanism for photos
After a bit of a break from the project, the new
update is finally out! This is the continuation
step for creating a responsive desktop version of
the app.

- Created a photo grid container that adjusts to
the user device width - for narrow devices, grid
does allow 1 column layout, with the maximum limit
of 3 columns for huge screens.

- Restricted the app size (beside the main Landing
Page) to a 1320px limit to ensure the app size
would not get out of control on huge screens.

- Created a general function that returns style
classes that are common for panels. This will
reduce redundancy between all the panel files.