fix: repair infected panel positioning issue
Solved an issue, where the bottom panel (that was
supposed to be infected into the top panel) got
set outside of the Teleport target div element.
This problem occured when the page content was
lower than the actual device height. Due to this
occurence, panels could not work as intended.

This issue, among some other related problems are
now fixed.