chore: add load when accessing detailed photo page
A small little fix, that enhances the user
experience. Now, when user is accessing detailed
photo page, instead of a sudden content shift
being seen, there is a nice little animation
that hides the process of layout shift until
the photo is loaded (that is, right after layout
has been shifted and remains still).